move `graph::converse` to new relation module...
... and make public.

This is a generic method on IndexedCoproducts thought of as encoding
relations, and was used in some downstream projects.
