+++ src/strict/eval.rs
@@ -5,9 +5,9 @@ use crate::finite_function::*;
 use crate::indexed_coproduct::*;
 use crate::semifinite::*;
 
-use crate::strict::graph::converse;
 use crate::strict::layer::layer;
 use crate::strict::open_hypergraph::*;
+use crate::strict::relation::converse;
 
 use num_traits::Zero;
 use std::default::Default;
